requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
flask>=3.0.0
